Add sfsu scoop package and use for some subcommands
Use sfsu for 'list' and 'info' subcommands. I'm not sure if 'depends' works
properly, so I have left it out. 'sfsu list' has the advantage of not listing
every package twice when installing for all users.

NOTE: 'sfsu list' is deprecated and 'sfsu search --installed' will replace it,
so I am not sure if I will have to replace scoop-search.exe when this happens. I
would rather use that for search.