Set scheduled task to set filetype on login using
The previous method using dism.exe is the official way from Microsoft
wasn't working for me and it seems to be because it requires the entire
filetypes list to be set or it gets rejected. It's also system-wide
rather than in the user scope, which isn't such a problem for me but it
still sucks.

The SetUserFTA.exe tool I am now using instead doesn't have these
disadvantages and using a simple list for the associations means that
it can easily be updated.

http://kolbi.cz/blog/?p=346
SetUserFTA.exe 1.8.2
ProgIDTool.exe 1.1